Исправить UPP-07..10: код ни разу не собирался в UE
Задачи UPP-07..10 были помечены [x], но GV2Editor не собирался ни
разу — правило плана «тест не отмечается зелёным без реальной
проверки» было нарушено на всех четырёх. RunUBT.sh GV2Editor выявил:

- GV2PreparedUiValue.h: FGV2PreparedUiObject хранил FGV2PreparedUiValue
  по значению до её объявления (циклическая неполнота типа). Класс
  переставлен — значение объявляется первым, контейнеры после;
  AsObject()/AsArray() вынесены в .cpp.
- GV2UiCapability.cpp/GV2UiMutationPlan.cpp: написаны под вымышленную
  форму GV2ContentCore::FCompiledUiFieldSpec (EUiFieldKind::Bool/
  Integer/Number/String, плоские TargetKind/IntMin/NumberMin/
  ObjectFields), не совпадающую с реальным API после UPP-02..04
  (Scalar-обёртка, RefTargetKind, Fields из FCompiledUiObjectField).
  Переписаны под реальный API.
- GV2PropertyConsumers.cpp: кастовал к несуществующему
  IGV2UiInteractionEmitter::SetBindingHandle. Добавлен реальный
  UGV2UiBindingTarget/IGV2UiBindingTarget; также
  EGV2PrimitiveScalePolicy::Fill (не существует) → PreserveAspect.
- UiSchema.h (GV2ContentCore): отсутствовал #include <map>,
  маскировался транзитивным include в portable-сборке.
- Тестовые файлы UPP-07..10 переписаны под реальный API; тест Prepare
  Purity Check передавал nullptr вместо host-виджета и ожидал
  успешного Prepare, что противоречит собственному Done-пункту
  («fallible-работа... проверка target») — заменён на реальный
  UGV2PanelWidgetBase с именованными Label/Bar; assertion числа
  mutations исправлен (2, не 3 — full-field семантика не мутирует
  capability без схемы и без прежнего committed-значения).

Заодно исправлены три не связанных с UPP предсуществующих ошибки
компиляции, блокировавших любую сборку GV2Editor: GetDesiredViewModel
(не существует, GetTextContent), FGV2UiBindingHandle{1,1} (нет такого
конструктора, ::Create), printf с лишним аргументом в
GV2WidgetSemanticFontSizeContractTests.cpp.

Сборка GV2Editor и automation-тесты GV2.UI.* +
GV2.Runtime.ContentCore.UiSchema подтверждены зелёными через
UnrealEditor-Cmd. Полный прогон GV2.* выявил 12 не связанных с этим
планом предсуществующих красных тестов (Location-композиты,
CoreRepeater, RichText style asset) — вне рамок этой задачи, не
трогались.

+++ Source/GV2/Private/Tests/GV2PreparedUiValueTests.cpp
@@ -97,8 +97,8 @@ bool FGV2PreparedUiValueTest::RunTest(const FString& Parameters)
 
         const TSharedRef<const FGV2PreparedUiArray> Arr = FGV2PreparedUiArray::Create(MoveTemp(Items));
         TestEqual(TEXT("Array count"), Arr->Num(), 2);
-        TestEqual(TEXT("Array item 0"), (*Arr)[0].AsInteger(), 100);
-        TestEqual(TEXT("Array item 1"), (*Arr)[1].AsInteger(), 200);
+        TestEqual(TEXT("Array item 0"), (*Arr)[0].AsInteger(), int64(100));
+        TestEqual(TEXT("Array item 1"), (*Arr)[1].AsInteger(), int64(200));
 
         const FGV2PreparedUiValue ArrVal = FGV2PreparedUiValue::MakeArray(Arr);
         TestTrue(TEXT("Array value kind"), ArrVal.IsArray());
